refactor: migrate protocol execution logic
Alice and Bob logic is migrated into protocol module.
